+++ euphonic/cli/intensity_map.py
@@ -23,7 +23,7 @@ def main(params: List[str] = None) -> None:
     q_spacing = _get_q_distance(args.length_unit, args.q_spacing)
     recip_length_unit = q_spacing.units
 
-    frequencies_only = (args.weights != 'coherent')
+    frequencies_only = (args.weighting != 'coherent')
 
     if isinstance(data, euphonic.ForceConstants):
         print("Force Constants data was loaded. Getting band path...")
@@ -38,13 +38,12 @@ def main(params: List[str] = None) -> None:
         x_tick_labels = get_qpoint_labels(modes.qpts,
                                           cell=modes.crystal.to_spglib_cell())
     modes.frequencies_unit = args.energy_unit
+    ebins = _get_energy_bins(modes, args.ebins + 1, emin=args.e_min,
+                             emax=args.e_max)
 
     print("Computing intensities and generating 2D maps")
 
-    if args.weights.lower() == 'coherent':
-        # calculate_sqw_map uses bin edges
-        ebins = _get_energy_bins(
-            modes, args.ebins + 1, emin=args.e_min, emax=args.e_max)
+    if args.weighting.lower() == 'coherent':
         if args.temperature is not None:
             if not isinstance(data, euphonic.ForceConstants):
                 raise TypeError("Cannot generate Debye-Waller factor without "
@@ -64,11 +63,8 @@ def main(params: List[str] = None) -> None:
         spectrum = (modes.calculate_structure_factor(dw=dw)
                     .calculate_sqw_map(ebins))
 
-    elif args.weights.lower() == 'dos':
-        # calculate_dos_map uses bin centres
-        ebins = _get_energy_bins(
-            modes, args.ebins, emin=args.e_min, emax=args.e_max)
-        spectrum = calculate_dos_map(modes, ebins)
+    elif args.weighting.lower() == 'dos':
+        spectrum = modes.calculate_dos_map(ebins)
 
     if args.q_broadening or args.energy_broadening:
         spectrum = spectrum.broaden(
@@ -104,33 +100,19 @@ def main(params: List[str] = None) -> None:
     matplotlib_save_or_show(save_filename=args.save_to)
 
 
-def calculate_dos_map(modes: euphonic.QpointPhononModes,
-                      ebins: euphonic.Quantity) -> euphonic.Spectrum2D:
-    from euphonic.util import _calc_abscissa
-    q_bins = _calc_abscissa(modes.crystal.reciprocal_cell(), modes.qpts)
-
-    bin_indices = np.digitize(modes.frequencies.magnitude, ebins.magnitude)
-    intensity_map = np.zeros((modes.n_qpts, len(ebins) + 1))
-    first_index = np.tile(range(modes.n_qpts),
-                          (3 * modes.crystal.n_atoms, 1)).transpose()
-    np.add.at(intensity_map, (first_index, bin_indices), 1)
-
-    return euphonic.Spectrum2D(q_bins, ebins,
-                               intensity_map[:, :-1] * ureg('dimensionless'))
-
-
 def get_parser() -> argparse.ArgumentParser:
-    parser, sections = _get_cli_parser(features={'read-fc', 'read-modes',
-                                                 'q-e', 'map', 'btol', 'ebins',
-                                                 'weights', 'plotting'})
+    parser, sections = _get_cli_parser(
+        features={'read-fc', 'read-modes', 'q-e', 'map', 'btol', 'ebins',
+                  'ins-weighting', 'plotting'})
     parser.description = (
         'Plots a 2D intensity map from the file provided. If a force '
         'constants file is provided, a band structure path is '
         'generated using Seekpath')
 
     sections['q'].description = (
-        '"GRID" options relate to Monkhorst-Pack sampling for the Debye-Waller'
-        ' factor, and only apply when --weights=coherent and --temperature is '
-        'set. "Q" options relate to the x-axis of spectrum data.')
+        '"GRID" options relate to Monkhorst-Pack sampling for the '
+        'Debye-Waller factor, and only apply when --weighting=coherent '
+        'and --temperature is set. "Q" options relate to the x-axis of '
+        'spectrum data.')
 
     return parser