+++ tests_and_analysis/test/script_tests/test_powder_map.py
@@ -20,17 +20,19 @@ powder_map_output_file = os.path.join(get_script_test_data_path(),
 quick_calc_params = ['--npts=10', '--npts-min=10', '--q-spacing=1']
 powder_map_params = [
     [nacl_prim_fc_file],
-    [nacl_prim_fc_file, *quick_calc_params],
     [nacl_prim_fc_file, '--temperature=1000', *quick_calc_params],
     [nacl_prim_fc_file, '--temperature=1000', '--weights=coherent',
      *quick_calc_params],
+    [nacl_prim_fc_file, '--temperature=1000', '--weighting=coherent',
+     *quick_calc_params],
     [nacl_prim_fc_file, '--v-min=0', '--v-max=1e-10', *quick_calc_params],
     [nacl_prim_fc_file, '--energy-unit=meV', *quick_calc_params],
-    [nacl_prim_fc_file, '--weights=coherent', '--cmap=bone'],
+    [nacl_prim_fc_file, '--weighting=coherent', '--cmap=bone',
+     *quick_calc_params],
     [graphite_fc_file, '-w', 'dos', '--y-label=DOS', '--title=DOS TITLE',
      *quick_calc_params],
     [graphite_fc_file, '--e-min=50', '-u=cm^-1', '--x-label=wavenumber',
-     *quick_calc_params],
+     *quick_calc_params, '-w=coherent-dos'],
     [graphite_fc_file, '--e-min=-100', '--e-max=1000', '--ebins=100',
      '--energy-unit=cm^-1', *quick_calc_params],
     [graphite_fc_file, '--energy-broadening=2e-3', '-u=eV',
@@ -42,7 +44,10 @@ powder_map_params = [
     [graphite_fc_file, '--asr', *quick_calc_params],
     [graphite_fc_file, '--asr=realspace', '--dipole-parameter=0.75',
      *quick_calc_params],
-    [nacl_prim_fc_file, *quick_calc_params]]
+    [nacl_prim_fc_file, '-w=incoherent-dos', '--pdos=Na',
+     *quick_calc_params],
+    [nacl_prim_fc_file, '-w=coherent-plus-incoherent-dos', '--pdos=Cl',
+     *quick_calc_params]]
 
 
 @pytest.mark.integration
@@ -67,9 +72,10 @@ class TestRegression:
         image_data = get_current_plot_image_data()
 
         with open(powder_map_output_file, 'r') as expected_data_file:
-            expected_data_key = args_to_key(powder_map_args)
-            expected_image_data = json.load(
-                expected_data_file)[expected_data_key]
+            # Test deprecated --weights until it is removed
+            key = args_to_key(powder_map_args).replace(
+                    'weights', 'weighting')
+            expected_image_data = json.load(expected_data_file)[key]
         for key, value in image_data.items():
             if key == 'extent':
                 # Lower bound of y-data (energy) varies by up to ~2e-6 on
@@ -101,13 +107,44 @@ class TestRegression:
             euphonic.cli.powder_map.main(powder_map_args)
 
     @pytest.mark.parametrize('powder_map_args', [
+        [nacl_prim_fc_file, '--weights=dos']])
+    def test_weights_emits_deprecation_warning(
+            self, inject_mocks, powder_map_args):
+        with pytest.warns(DeprecationWarning):
+            euphonic.cli.powder_map.main(powder_map_args + quick_calc_params)
+
+    @pytest.mark.parametrize('powder_map_args', [
         [nacl_prim_fc_file, '-w=incoherent']])
-    def test_invalid_weights_raises_causes_exit(self, powder_map_args):
+    def test_invalid_weighting_raises_causes_exit(self, powder_map_args):
         # Argparse should call sys.exit on invalid choices
         with pytest.raises(SystemExit) as err:
             euphonic.cli.powder_map.main(powder_map_args)
         assert err.type == SystemExit
         assert err.value.code == 2
+
+    @pytest.mark.parametrize('powder_map_args', [
+        [nacl_prim_fc_file, '-w=coherent', '--pdos', 'Na']])
+    def test_coherent_weighting_and_pdos_raises_value_error(
+            self, powder_map_args):
+        with pytest.raises(ValueError):
+            euphonic.cli.powder_map.main(powder_map_args + quick_calc_params)
+
+    @pytest.mark.parametrize('powder_map_args', [
+        [nacl_prim_fc_file, '--pdos']])
+    def test_no_pdos_args_raises_causes_exit(self, powder_map_args):
+        with pytest.raises(SystemExit) as err:
+            euphonic.cli.powder_map.main(powder_map_args)
+        assert err.type == SystemExit
+        assert err.value.code == 2
+
+    @pytest.mark.parametrize('powder_map_args', [
+        [nacl_prim_fc_file, '--pdos', 'Na', 'Cl']])
+    def test_multiple_pdos_args_raises_causes_exit(self, powder_map_args):
+        with pytest.raises(SystemExit) as err:
+            euphonic.cli.powder_map.main(powder_map_args)
+        assert err.type == SystemExit
+        assert err.value.code == 2
+
 
 
 @patch('matplotlib.pyplot.show')